refactor: extract search post filters
